64-bit cleanness: lots more replacing of UInt with UWord as necessary.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@2920

+++ coregrind/vg_memory.c
@@ -603,7 +603,7 @@ void VG_(unpad_address_space)(void)
 
    while (s && addr <= VG_(valgrind_last)) {
       if (addr < s->addr) {
-         ret = VG_(do_syscall)(__NR_munmap, (UInt)addr, s->addr - addr);
+         ret = VG_(do_syscall)(__NR_munmap, addr, s->addr - addr);
       }
          
       addr = s->addr + s->len;
